% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ADMM}
\alias{ADMM}
\title{ADMM class.}
\description{
The attributes include regression data(dataX, dataY, lambda, rho), iteration parameters(max_iteration, eps_abs, eps_rel), variables to be solved(x, z, u), etc;
The function includes initialization of the class(ADMM()), determination of convergence(ifConvergent()), updates of each solving variable(UpdateX(), UpdateZ(), UpdateU()), and complete solution of the function(getResult()).
}
