% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CG}
\alias{CG}
\title{CG class.}
\description{
The function includes initialization of the class(CG()), calculate the loss of beta_k,two ways to determine the step(alpha),the solving function of minimize loss by iteration,and complete solution of the function(getResult())
}
