% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.R
\name{getData}
\alias{getData}
\title{Get simulation data}
\usage{
getData(seed_num, n, p)
}
\arguments{
\item{seed_num}{the random seed}

\item{n}{the sample size}

\item{p}{the feature size}
}
\value{
the generated data and beta as a list
}
\description{
Get simulation data
}
