% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{L_BFGS}
\alias{L_BFGS}
\title{class L_BFGS: using L-BFGS algorithm to solve the Optimization Problems
              where the target_objection is ridge regression.}
\description{
The function includes initialization of the class(L-BFGS()), 
the ridge regression loss function(ridge_objective()),
function to monitor the progress of the optimization process(monitorProgress()), 
the solving function of minimize loss by iteration(monitorProgress()),
and complete solution of the function by my_lbfgs::lbfgs_optimize(getResult())
}
