% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.R
\name{plotSample}
\alias{plotSample}
\title{Draw contour maps(when the dimensions of beta is 2).}
\usage{
plotSample(X, y, lambda, beta)
}
\arguments{
\item{X}{the independent variable X (size n×p)}

\item{y}{the response variable y (size n×1)}

\item{lambda}{the penalty factor (size 1×1)}

\item{beta}{the estimated value of beta (size p×1)}
}
\value{
the contour map
}
\description{
Draw contour maps(when the dimensions of beta is 2).
}
\examples{
 
library(ggplot2)
library(ggisoband)
data = getData(123,100,2)
result = RidgeRegression(data$X,data$y,lambda = 0.1,max_iter = 100)
plotSample(data$X,data$y,lambda = 0.1,beta = result$betaHis)
}
