% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RidgeRegression_CG}
\alias{RidgeRegression_CG}
\title{Compute the optimal estimates of beta by minimize the loss function with l2 penalty through CG method}
\usage{
RidgeRegression_CG(X, y, lambda, t = 0.1, eps = 1e-05, max_iteration = 1000L)
}
\arguments{
\item{X}{the independent variable matrix}

\item{y}{the dependent variable vector}

\item{lambda}{the coeffient of l2 regularization}
}
\value{
the list contains the estimated value of beta,the matrix of the beta under every iteration,the iterations number,the final loss
}
\description{
Compute the optimal estimates of beta by minimize the loss function with l2 penalty through CG method
}
\examples{
seed_num = 123
n = 100
p = 2
simudata = getData(seed_num,n,p)
result = RidgeRegression_CG(simudata$X,simudata$y,lambda = 0.1,max_iteration = 1000)
result$betaHat
}
