% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RidgeRegression_LBFGS}
\alias{RidgeRegression_LBFGS}
\title{Solve ridge regression using L-BFGS.}
\usage{
RidgeRegression_LBFGS(
  X,
  y,
  lambda,
  mem_size = 8L,
  max_iterations = 64L,
  g_epsilon = 1e-05,
  past = 3L,
  delta = 1e-06
)
}
\arguments{
\item{X}{the independent variable X (size n×p)}

\item{y}{the response variable y (size n×1)}

\item{lambda}{the penalty factor (size 1×1)}

\item{max_iterations}{the maximum number of iterations}

\item{g_epsilon}{the epsilon for grad convergence test}

\item{past}{the distance for delta-based convergence test}

\item{delta}{for convergence test}

\item{men_size}{the number of corrections to approximate the inverse hessian matrix}
}
\value{
Estimated value of targetValues, beta(last and historical), iterations and loss
}
\description{
Solve ridge regression using L-BFGS.
}
