\name{RRopt-package}
\alias{RRopt-package}
\alias{RRopt}
\docType{package}
\title{
\packageTitle{RRopt}
}
\description{
\packageDescription{RRopt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RRopt}
\packageIndices{RRopt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RRopt}

Maintainer: \packageMaintainer{RRopt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
